bugfix: constrainHud feature
Set constrainHud feature to be false by default, the user needs to
manually decide whether or not they want certain features to be enabled
or not. Features by default should always be false while fixes always
be true.

	modified:   GodEater1-2Fix.yml

Originally the mod was designed and solely tested for 32:9, thinking
that if it works for that aspect ratio it would work for 21:9 as well.
Turns out that was not the case as the calculation for constraining
the HUD was more complex than 32:9 made it out to be. Implemented a new
calculation which will constrain the HUD properly to center and be 16:9.

	modified:   src/dllmain.cpp